[TIMOB-18382] TitaniumKit should have 0 Warnings on OS X

+++ Source/TitaniumKit/src/UI/ImageView.cpp
@@ -141,7 +141,7 @@ namespace Titanium
 			
 			std::vector<std::string> images;
 			const auto item_count = object.GetPropertyNames().GetCount();
-			for (size_t i = 0; i < item_count; ++i) {
+			for (uint32_t i = 0; i < item_count; ++i) {
 				JSValue item = object.GetProperty(i);
 				TITANIUM_ASSERT(item.IsString());
 				images.push_back(static_cast<std::string>(item));
